Moved files to more appropriate folders.
