+++ LTU.SearchEngine.Backend/LTU.SearchEngine.Backend.Core/Model/ValueObjects/QueryNodes/TermNode.cs
@@ -1,4 +1,6 @@
-﻿namespace LTU.SearchEngine.Backend.Core.Model.ValueObjects.QueryNodes;
+﻿using LTU.SearchEngine.Backend.Core.SearchQueryBuilder;
+
+namespace LTU.SearchEngine.Backend.Core.Model.ValueObjects.QueryNodes;
 
 
 /// <summary>
