+++ LTU.SearchEngine.Backend/LTU.SearchEngine.Test/QueryParsing.Tests/QueryNode.Tests/RequiredNodeTests.cs
@@ -1,5 +1,6 @@
-﻿using LTU.SearchEngine.Backend.Core;
-using LTU.SearchEngine.Backend.Core.Model.ValueObjects.QueryNodes;
+﻿using LTU.SearchEngine.Backend.Core.Model.ValueObjects.QueryNodes;
+using LTU.SearchEngine.Backend.Core.SearchQueryBuilder;
+using LTU.SearchEngine.Backend.Core.SearchQueryBuilder.QueryNodes;
 using Moq;
 
 namespace LTU.SearchEngine.Test.QueryParsing.Tests.QueryNode.Tests;
